+++ Assets/EntitySystem2D/Runtime/EntitySystem2D/Components/EntityMyAnimator.cs
@@ -10,11 +10,11 @@ namespace GoldSprite.EntitySystem2D {
     public class EntityMyAnimator : MyAnimator{
         public void InitMetas(List<BehaviourMeta> behaviourMetaList)
         {
-            var dict = new Dictionary<Enum, AnimationClip>();
-            foreach(var meta in behaviourMetaList) {
-                dict[meta.BId] = meta.AnimClip;
-            }
-            SetAnimClips(dict);
+            //var dict = new Dictionary<Enum, AnimationClip>();
+            //foreach(var meta in behaviourMetaList) {
+            //    dict[meta.BId] = meta.AnimClip;
+            //}
+            //SetAnimClips(dict);
         }
     }
 }