+++ Assets/EntitySystem2D/Runtime/EntitySystem2D/Metas/AttackBehaviourMeta.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace GoldSprite.EntitySystem2D {
     [CreateAssetMenu]
-    public class MoveBehaviourMeta : EntityBehaviourMeta {
-        public override Type StateType => typeof(MoveState);
+    public class AttackBehaviourMeta : EntityBehaviourMeta {
+        public override Type StateType => typeof(AttackState);
     }
 }