+++ Assets/EntitySystem2D/Runtime/EntitySystem2D/Metas/BehaviourMeta.cs
@@ -9,22 +9,12 @@ using UnityEngine;
 
 namespace GoldSprite.EntitySystem2D {
     public abstract class BehaviourMeta : SerializedScriptableObject, IBehaviourMeta {
-        public virtual Enum BId {get;}
         public int priority;
         public int Priority => priority;
         public AnimationClip animClip;
         public virtual AnimationClip AnimClip => animClip;
         public virtual Type StateType { get; }
-        public virtual IBaseProps Props { get; }
     }
     public abstract class EntityBehaviourMeta : BehaviourMeta {
-        public EntityBId bid;
-        public override Enum BId => bid;
-    }
-
-
-    public enum EntityBId
-    {
-        Idle, Move
     }
 }