+++ Assets/EntitySystem2D/Runtime/EntitySystem2D/Metas/IBehaviourMeta.cs
@@ -6,11 +6,8 @@ using UnityEngine;
 
 namespace GoldSprite.EntitySystem2D {
     public interface IBehaviourMeta {
-
-        public Enum BId { get; }
         public int Priority { get; }
         public Type StateType { get; }
         public AnimationClip AnimClip { get; }
-        public IBaseProps Props { get; }
     }
 }