+++ Assets/EntitySystem2D/Runtime/EntitySystem2D/Metas/IdleBehaviourMeta.cs
@@ -6,7 +6,5 @@ namespace GoldSprite.EntitySystem2D {
     [CreateAssetMenu]
     public class IdleBehaviourMeta : EntityBehaviourMeta {
         public override Type StateType=> typeof(IdleState);
-        public BasePropsMeta props;
-        public override IBaseProps Props=> props;
     }
 }