+++ Assets/EntitySystem2D/Runtime/EntitySystem2D/Metas/PropsMetas/BasePropsMeta2.cs
@@ -1,7 +1,0 @@
-﻿using UnityEngine;
-
-namespace GoldSprite.EntitySystem2D {
-    [CreateAssetMenu]
-    public class BasePropsMeta2 : BasePropsMeta, IBaseProps2 {
-    }
-}