+++ Assets/EntitySystem2D/Runtime/EntitySystem2D/Metas/PropsMetas/EntityPropsMeta.cs
@@ -1,0 +1,20 @@
+﻿using GoldSprite.UFsm;
+using UnityEngine;
+
+namespace GoldSprite.EntitySystem2D {
+    //[CreateAssetMenu]
+    //public class EntityPropsMeta : PropsMeta, IEntityProps {
+    //    [SerializeField] private new string name;
+    //    public string Name { get => name; set => name = value; }
+    //    [SerializeField] private Vector2 direction;
+    //    public Vector2 Direction { get => direction; set => direction = value.magnitude > 1 ? value.normalized : value; }
+    //    public Vector2 velocity;
+    //    public Vector2 Velocity { get => velocity; set => velocity = value; }
+    //    public bool IsGround { get; } = true;
+    //    [SerializeField] private string speed;
+    //    public float Speed { get; set; } = 1;
+    //    public int Face { get; set; }
+    //    public bool AttackKey { get; set; }
+    //    public bool HurtKey { get; set; }
+    //}
+}