+++ Assets/Packages/UnityPlugins/MyAnimator/Runtime/MyAnimatorSystem/MyAnimator.cs
@@ -12,17 +12,16 @@ namespace GoldSprite.UnityPlugins.MyAnimator {
         [ManualRequire]
         public Animator anims;
         [OdinSerialize]
-        private readonly Dictionary<Enum, AnimationClip> animClips = new();
+        private readonly Dictionary<string, AnimationClip> animClips = new();
 
         //引用
         [Header("动画事件")]
-        public Action<Enum> OnAnimEndEvent = (p) => { };
-        public Action<Enum, Enum> OnAnimTranEvent = (p, p2) => { };  //last, next
+        public Action<string> OnAnimEndEvent = (p) => { };
+        public Action<string, string> OnAnimTranEvent = (p, p2) => { };  //last, next
         [Header("实时")]
         public string CAnimName = "";
         public string LastAnimName = "";
         public string LastAnimEndName = "";
-        public Enum CAnimEnum, LastAnimEnum, LastAnimEndEnum;
         public AnimationClip CAnimClip => anims.GetCurrentAnimatorClipInfo(0)[0].clip;
         public AnimatorStateInfo CAnimState => anims.GetCurrentAnimatorStateInfo(0);
         public float CAnimNormalizedTime;
@@ -31,7 +30,7 @@ namespace GoldSprite.UnityPlugins.MyAnimator {
         public bool CAnimTranslationing;
 
 
-        public void SetAnimClips<T>(Dictionary<T, AnimationClip> clips) where T:Enum
+        public void SetAnimClips(Dictionary<string, AnimationClip> clips)
         {
             animClips.Clear();
             foreach (var (k, v) in clips) {
@@ -43,50 +42,37 @@ namespace GoldSprite.UnityPlugins.MyAnimator {
         {
             try {
                 CAnimName = CAnimClip.name;
-                CAnimEnum = GetClipKey(CAnimClip);
                 CAnimNormalizedTime = CAnimState.normalizedTime;
                 CAnimClipLooping = CAnimClip.isLooping;
                 CAnimStateLoop = CAnimState.loop;
                 CAnimTranslationing = anims.IsInTransition(0);
 
-                if (LastAnimEndName != CAnimName && IsCurrentAnimEnd(CAnimName)) {
-                    LastAnimEndName = CAnimName;
-                    //PlayAnim(CAnimName, 0, CAnimState.normalizedTime - 1);
-                    LogTool.NLog("MyAnimator", $"动画播放结束事件: {LastAnimName}");
-                    //CurrentAnimEndAction?.Invoke(LastAnimName);
-                    var clipKey = GetClipKey(CAnimClip);
-                    OnAnimEndEvent?.Invoke(clipKey);
-                }
                 if (LastAnimName != CAnimName) {
                     LastAnimEndName = "";
                     LogTool.NLog("MyAnimator", $"动画转换事件{LastAnimName} -> {CAnimName}");
-                    OnAnimTranEvent?.Invoke(LastAnimEnum, CAnimEnum);
+                    OnAnimTranEvent?.Invoke(LastAnimName, CAnimName);
                     LastAnimName = CAnimName;
-                    LastAnimEnum = CAnimEnum;
+                }
+
+                if (LastAnimEndName != CAnimName && IsCurrentAnimEnd(CAnimName)) {
+                    //Play(CAnimName, 0, CAnimState.normalizedTime - 1);
+                    LogTool.NLog("MyAnimator", $"动画播放结束事件: {CAnimName}");
+                    OnAnimEndEvent?.Invoke(CAnimName);
+                    LastAnimEndName = CAnimName;
                 }
             }
             catch (Exception) { }
         }
 
 
-        public Enum GetClipKey(AnimationClip clip)
+        public string GetClipKey(AnimationClip clip)
         {
             return animClips.FirstOrDefault(p => p.Value == clip).Key;
         }
 
-        public void Play(Enum key, int layer = 0, float normalizedTime = 0)
-        {
-            anims.Play(animClips[key].name, layer, normalizedTime);
-        }
-        public void Play(Enum key)
-        {
-            anims.Play(animClips[key].name);
-        }
-
-        public void PlayAnim(string animName, int layer = 0, float normalizedTime = 0)
-        {
-            anims.Play(animName, layer, normalizedTime);
-        }
+        public void Play(string animName) => anims.Play(animName);
+        public void Play(string animName, int layer = 0) => anims.Play(animName, layer);
+        public void Play(string animName, int layer = 0, float normalizedTime = 0) => anims.Play(animName, layer, normalizedTime);
 
         public bool IsAnimName(string animName)
         {