+++ Assets/Packages/UnityPlugins/UFsm/Scripts/Runtime/Fsm/Basics/States/DeathState.cs
@@ -1,0 +1,34 @@
+
+using UnityEngine;
+
+namespace GoldSprite.UFsm {
+    public class DeathState : BaseState {
+        public override bool CanTranSelf { get; protected set; } = false;
+        public DeathState(BaseFsm fsm) : base(fsm)
+        {
+        }
+
+        public override bool Enter()
+        {
+            return Props.DeathKey;
+        }
+        public override bool Exit()
+        {
+            return !Props.DeathKey;
+        }
+
+        public override void OnEnter()
+        {
+            //¶¯»­
+            Fsm.AnimCtrls.Play(AnimName, 0, 0);
+        }
+
+        public override void OnExit()
+        {
+        }
+
+        public override void FixedUpdate()
+        {
+        }
+    }
+}