+++ Assets/Packages/UnityPlugins/UFsm/Scripts/Runtime/Fsm/Interfaces/Fsm.cs
@@ -3,6 +3,7 @@ using System;
 using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.Networking.Types;
+using static UnityEngine.GraphicsBuffer;
 
 namespace GoldSprite.UFsm {
     public class Fsm : SerializedMonoBehaviour, IFsm {
@@ -39,7 +40,7 @@ namespace GoldSprite.UFsm {
             foreach (var state in states.Values) {
                 if (EnterState(state, targetState)) targetState = state;
             }
-            var change = targetState != CState;
+            var change = targetState != CState || (targetState.CanTranSelf && targetState.Enter());
             if (!change && CState.Exit()) { targetState = DefaultState; change = true; }
             if (change) OnEnterState(targetState);
             return change;